ref: fixed some bugs

+++ src/main/java/uz/ccrew/trainerworkloadservice/service/impl/TrainerWorkloadServiceImpl.java
@@ -1,5 +1,6 @@
 package uz.ccrew.trainerworkloadservice.service.impl;
 
+import uz.ccrew.trainerworkloadservice.enums.ActionType;
 import uz.ccrew.trainerworkloadservice.dto.summary.YearsDTO;
 import uz.ccrew.trainerworkloadservice.dto.summary.MonthsDTO;
 import uz.ccrew.trainerworkloadservice.entity.TrainerWorkload;
@@ -8,6 +9,7 @@ import uz.ccrew.trainerworkloadservice.service.TrainerWorkloadService;
 import uz.ccrew.trainerworkloadservice.repository.TrainerWorkLoadRepository;
 import uz.ccrew.trainerworkloadservice.dto.summary.TrainerMonthlySummaryDTO;
 import uz.ccrew.trainerworkloadservice.dto.trainer_workload.TrainerWorkloadDTO;
+import uz.ccrew.trainerworkloadservice.dto.trainer_workload.TrainerMonthlySummaryProjection;
 
 import lombok.extern.slf4j.Slf4j;
 import lombok.RequiredArgsConstructor;
@@ -16,7 +18,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.*;
 import java.time.Month;
-import java.time.format.TextStyle;
 
 @Slf4j
 @Service
@@ -31,62 +32,45 @@ public class TrainerWorkloadServiceImpl implements TrainerWorkloadService {
             throw new ObjectNotProvidedException("ActionType cannot be null");
         }
 
-        switch (dto.getActionType()) {
-            case ADD -> {
-                TrainerWorkload newWorkload = TrainerWorkload.builder()
-                        .trainerUsername(dto.getTrainerUsername())
-                        .trainerLastName(dto.getTrainerLastName())
-                        .trainerFirstName(dto.getTrainerFirstName())
-                        .trainingDuration(dto.getTrainingDuration())
-                        .trainingDate(dto.getTrainingDate())
-                        .isActive(dto.isActive())
-                        .build();
-                trainerWorkLoadRepository.save(newWorkload);
-                log.info("Added new trainer workload for {} on {}", dto.getTrainerUsername(), dto.getTrainingDate());
-            }
-
-            case DELETE -> {
-                trainerWorkLoadRepository.deleteByTrainerUsernameAndTrainingDate(dto.getTrainerUsername(), dto.getTrainingDate());
-                log.info("Deleted the trainer workload for {} on {}", dto.getTrainerUsername(), dto.getTrainingDate());
-            }
+        if (dto.getActionType() == ActionType.ADD) {
+            TrainerWorkload newWorkload = TrainerWorkload.builder()
+                    .trainerUsername(dto.getTrainerUsername())
+                    .trainerLastName(dto.getTrainerLastName())
+                    .trainerFirstName(dto.getTrainerFirstName())
+                    .trainingDuration(dto.getTrainingDuration())
+                    .trainingDate(dto.getTrainingDate())
+                    .isActive(dto.isActive())
+                    .build();
+            trainerWorkLoadRepository.save(newWorkload);
+            log.info("Added new trainer workload for {} on {}", dto.getTrainerUsername(), dto.getTrainingDate());
+        } else if (dto.getActionType() == ActionType.DELETE) {
+            trainerWorkLoadRepository.deleteByTrainerUsernameAndTrainingDate(dto.getTrainerUsername(), dto.getTrainingDate());
+            log.info("Deleted the trainer workload for {} on {}", dto.getTrainerUsername(), dto.getTrainingDate());
         }
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<TrainerMonthlySummaryDTO> getMonthlyWorkload(String username) {
-        List<Object[]> rawData = trainerWorkLoadRepository.getMonthlySummary(username);
-        if (rawData.isEmpty()) {
+        List<TrainerMonthlySummaryProjection> workload = trainerWorkLoadRepository.getMonthlySummary(username);
+        if (workload.isEmpty()) {
             return List.of();
         }
-
         Map<Integer, List<MonthsDTO>> yearToMonthsMap = new LinkedHashMap<>();
-        String trainerUsername = null;
-        String trainerFirstName = null;
-        String trainerLastName = null;
-        boolean isActive = false;
+        TrainerMonthlySummaryProjection firstEntry = workload.getFirst();
 
-        for (Object[] row : rawData) {
-            trainerUsername = (String) row[0];
-            trainerFirstName = (String) row[1];
-            trainerLastName = (String) row[2];
-            isActive = (boolean) row[3];
-
-            int year = ((Number) row[4]).intValue();
-            int month = ((Number) row[5]).intValue();
-            String monthName = Month.of(month).getDisplayName(TextStyle.FULL, Locale.ENGLISH);
-            double totalDuration = ((Number) row[6]).doubleValue();
-
+        for (TrainerMonthlySummaryProjection projection : workload) {
             MonthsDTO monthsDTO = MonthsDTO.builder()
-                    .month(monthName)
-                    .totalDuration(totalDuration)
+                    .month(Month.of(projection.getMonth()))
+                    .totalDuration(projection.getTotalDuration())
                     .build();
 
             yearToMonthsMap
-                    .computeIfAbsent(year, y -> new ArrayList<>())
+                    .computeIfAbsent(projection.getYear(), y -> new ArrayList<>())
                     .add(monthsDTO);
         }
 
-        List<YearsDTO> yearsList = (List<YearsDTO>) yearToMonthsMap.entrySet().stream()
+        List<YearsDTO> yearsList = yearToMonthsMap.entrySet().stream()
                 .map(entry -> YearsDTO.builder()
                         .year(entry.getKey())
                         .months(entry.getValue())
@@ -94,12 +78,13 @@ public class TrainerWorkloadServiceImpl implements TrainerWorkloadService {
                 .toList();
 
         TrainerMonthlySummaryDTO summary = TrainerMonthlySummaryDTO.builder()
-                .trainerUsername(trainerUsername)
-                .trainerFirstName(trainerFirstName)
-                .trainerLastName(trainerLastName)
-                .isActive(isActive)
+                .trainerUsername(firstEntry.getTrainerUsername())
+                .trainerFirstName(firstEntry.getTrainerUsername())
+                .trainerLastName(firstEntry.getTrainerLastName())
+                .isActive(firstEntry.isActive())
                 .years(yearsList)
                 .build();
+
         return List.of(summary);
     }
 }
